メモ入力のタイマー処理の改善
