+++ src/app/(authenticated)/ThemeSelect/page.tsx
@@ -1,44 +1,17 @@
 "use client";
-import { useAtomValue, useSetAtom } from "jotai";
+import { useAtomValue } from "jotai";
 import { useRouter } from "next/navigation";
-import { useEffect, useState } from "react";
 
+// import { useEffect } from "react";
 import { Loader } from "@/component/Loader";
 import { useGetThemes } from "@/hooks/useGetThemes";
-import { countTheme, setThemeAtom } from "@/store/setting";
-import type { Theme } from "@/types/database";
-
-
+import { countTheme } from "@/store/setting";
 
 export default function ThemeSelectPage() {
   const ThemesToScribble = useAtomValue(countTheme); // 設定されたテーマ数を取得
-  const { error, loading, themes } = useGetThemes();
-  const setThemes = useSetAtom(setThemeAtom);
+  const { allThemes, error, loading, randomSelect } = useGetThemes();
+
   const router = useRouter();
-  const [selected, setSelected] = useState<Theme[]>([]);
-
-  useEffect(() => {
-    if (themes.length > 0) {
-      const selectedThemes = randomSelect(themes.slice(), ThemesToScribble);
-      setSelected(selectedThemes);
-    }
-  }, [themes, ThemesToScribble]);
-
-  useEffect(() => {
-    if (selected.length > 0) {
-      setThemes(selected);
-    }
-  }, [selected, setThemes]);
-
-  useEffect(() => {
-    const fetchThemes = async () => {
-      const response = await fetch("/api/themes");
-      const data = await response.json();
-      setThemes(data);
-    };
-
-    fetchThemes();
-  }, [setThemes]); // setThemesを依存配列に追加
 
   if (loading) {
     return <Loader />;
@@ -48,15 +21,7 @@ export default function ThemeSelectPage() {
     return <div>{error}</div>;
   }
 
-  function randomSelect(theme: Theme[], num: number): Theme[] {
-    const newTheme: Theme[] = [];
-    while (newTheme.length < num && theme.length > 0) {
-      const rand = Math.floor(Math.random() * theme.length);
-      newTheme.push(theme[rand]);
-      theme.splice(rand, 1);
-    }
-    return newTheme;
-  }
+  const selected = randomSelect(allThemes, ThemesToScribble);
 
   const handleStart = () => {
     router.push("/MemoEditor");
